fix: store & retrieve entire state (#1)
The current way of storing and retrieving data to and from the file is
error prone; and there is no indication as to which day it was entered
for and which day I missed.

The current implementation stores the file in the following way:
```
item1 item2 item3
X Y Y
Y Y X
Y Y Y
X X X
```

Clearly there is no indication of the day and if I miss to call `todo`
on any given day then there is no way we could add `?` in its place and
the entire system breaks if there are less number of lines than what is
expected (1 for each day of week and todo's line)

Storing the day along the line similar to the following could be a
natural idea
```
item1 item2 item3
Mon X Y Y
Wed Y Y X
Thu Y Y Y
Fri X X X
```

This could be done if I implement a parser for it; though it will be a
really simple one, but will increase the complexity of the project.

This commit tries to fix all of the problems by defining a `State` which
will store the todo's and the completion info for them for each day of
week and if the completion info is not found then it is easy to just add
`?` in its place.

I'm adding dependency for a simple project, yes; because this reduces
the complexity that I have to think of, serde will handle the storing of
the entire `State` and retrieving it.

`serde_yaml` is something that could change in future to different
formats like json or binary.